generalize Item Form functions for use in Edit Item Page
