+++ app/src/main/java/com/example/househomey/AddItemFragment.java
@@ -33,18 +33,15 @@ import java.util.Objects;
  *
  * @author Owen Cooke
  */
-public class AddItemFragment extends Fragment {
-    private Date dateAcquired;
-    private TextInputEditText dateTextView;
-    private final CollectionReference itemRef;
+public class AddItemFragment extends ItemFormFragment {
 
     /**
-     * Constructs a new AddItemFragment with a firestore reference
+     * Constructs a new AddItemFragment with a Firestore reference
      *
-     * @param itemRef A reference to a firestore collection of items
+     * @param itemRef A reference to a Firestore collection of items
      */
     public AddItemFragment(CollectionReference itemRef) {
-        this.itemRef = itemRef;
+        super(itemRef);
     }
 
     /**
@@ -61,34 +58,11 @@ public class AddItemFragment extends Fragment {
      */
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_add_item, container, false);
-
-        initDatePicker(rootView);
-        initTextValidators(rootView);
-
+        View rootView = super.onCreateView(inflater, container, savedInstanceState);
         // Add listener for confirm and back buttons
         rootView.findViewById(R.id.add_item_confirm_button).setOnClickListener(v -> addItem());
         rootView.findViewById(R.id.add_item_back_button).setOnClickListener(v -> navigateHomeWithIndicator((AppCompatActivity) getContext()));
-
         return rootView;
-    }
-
-    /**
-     * Checks whether the text input for a required form field is empty.
-     * Additionally, if empty, it sets an inline error message on the input field.
-     * If not empty, it removes the inline error.
-     *
-     * @param id Id of the required field's TextInputLayout
-     * @return a boolean indicating if the field is empty
-     */
-    private boolean isRequiredFieldEmpty(int id) {
-        TextInputLayout textInputLayout = getView().findViewById(id);
-        if (TextUtils.isEmpty(textInputLayout.getEditText().getText().toString().trim())) {
-            textInputLayout.setError("This field is required");
-            return true;
-        }
-        textInputLayout.setError(null);
-        return false;
     }
 
     /**
@@ -128,84 +102,5 @@ public class AddItemFragment extends Fragment {
                     Log.d("Firestore", "Failed to create new item");
                     getView().findViewById(R.id.add_item_error_msg).setVisibility(View.VISIBLE);
                 });
-    }
-
-    /**
-     * Gets the user input as a string from a given TextInputEditText
-     *
-     * @param id Id of the TextInputEditText object
-     * @return The user input String
-     * @throws NullPointerException if the input text is null
-     */
-    private String getInputText(int id) {
-        return Objects.requireNonNull(((TextInputEditText) requireView().findViewById(id)).getText()).toString();
-    }
-
-    /**
-     * Initializes and configures a Date Picker for selecting past/present acquisition dates.
-     *
-     * @param rootView The root view of the UI where the date picker is to be displayed.
-     */
-    private void initDatePicker(View rootView) {
-        dateTextView = rootView.findViewById(R.id.add_item_date);
-
-        // Create constraint to restrict dates to past/present
-        CalendarConstraints.Builder constraintsBuilder = new CalendarConstraints.Builder();
-        constraintsBuilder.setEnd(System.currentTimeMillis());
-        constraintsBuilder.setValidator(DateValidatorPointBackward.now());
-        MaterialDatePicker<Long> datePicker = MaterialDatePicker.Builder.datePicker()
-                .setInputMode(MaterialDatePicker.INPUT_MODE_CALENDAR)
-                .setCalendarConstraints(constraintsBuilder.build())
-                .build();
-        datePicker.addOnPositiveButtonClickListener(selection -> {
-            dateAcquired = new Date(selection);
-            dateTextView.setText(datePicker.getHeaderText());
-            ((TextInputLayout) rootView.findViewById(R.id.add_item_date_layout)).setError(null);
-        });
-
-        // Show DatePicker when date field selected
-        dateTextView.setOnClickListener(v -> datePicker.show(getParentFragmentManager(), "Date Picker"));
-    }
-
-    /**
-     * Initializes text validators for required input fields on the add item form.
-     *
-     * @param rootView The root view of the UI where the input fields are located.
-     */
-    private void initTextValidators(View rootView) {
-        TextInputEditText descView = rootView.findViewById(R.id.add_item_description);
-        TextInputEditText costView = rootView.findViewById(R.id.add_item_cost);
-
-        // Add text watchers for empty input to both description and cost
-        TextWatcher emptyTextWatcher = new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-            }
-
-            @Override
-            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-            }
-
-            @Override
-            public void afterTextChanged(Editable editable) {
-                if (editable.equals(descView.getEditableText())) {
-                    isRequiredFieldEmpty(R.id.add_item_description_layout);
-                } else if (editable == costView.getEditableText()) {
-                    isRequiredFieldEmpty(R.id.add_item_cost_layout);
-                }
-            }
-        };
-        descView.addTextChangedListener(emptyTextWatcher);
-        costView.addTextChangedListener(emptyTextWatcher);
-
-        // Add listener for rounding cost to 2 decimals
-        costView.setOnFocusChangeListener((v, b) -> {
-            String costString = costView.getText().toString();
-            try {
-                costString = new BigDecimal(costString).setScale(2, RoundingMode.HALF_UP).toString();
-                costView.setText(costString);
-            } catch (NumberFormatException ignored) {
-            }
-        });
     }
 }
